fix(Dockerfile): edit dockerfile to do a get and install
