+++ Dockerfile
@@ -1,13 +1,10 @@
 FROM golang:1.19
 
 WORKDIR /go/src/app
-
-# pre-copy/cache go.mod for pre-downloading dependencies and only redownloading them in subsequent builds if they change
-COPY go.mod go.sum ./
-RUN go mod download && go mod verify
-
 COPY . .
 
-RUN go build -v -o po-linter .
+RUN go get -d -v ./...
+RUN go install -v ./...
+RUN go build .
 
-CMD ["po-linter"]
+CMD [ "cloud-platform-po-linter" ]
