+++ cloud-platform-po-linter.yaml
@@ -1,0 +1,6 @@
+name: cloud-platform-po-linter
+description: Checks if Prometheus rule yaml has any errors
+
+runs:
+  using: 'docker'
+  image: 'Dockerfile'
