--- po-linter.yaml
@@ -1,6 +1,0 @@
-name: po-linter
-description: Checks if Prometheus rule yaml has any errors
-
-runs:
-  using: 'docker'
-  image: 'Dockerfile'
